+++ scripts/primitives/block3d.py
@@ -60,14 +60,14 @@ print(box_green.intersection_external_aabb_volume(box_blue, resolution))
 model = vm.core.VolumeModel([box, box_red, box_green, box_blue])
 model.babylonjs(debug=True)
 
-assert box.is_inside_shell(box_red, resolution) == True
-assert box_red.is_inside_shell(box, resolution) == False
+assert box.is_inside_shell(box_red) == True
+assert box_red.is_inside_shell(box) == False
 
-assert box.is_inside_shell(box_green, resolution) == False
-assert box_green.is_inside_shell(box, resolution) == False
+assert box.is_inside_shell(box_green) == False
+assert box_green.is_inside_shell(box) == False
 
-assert box.is_inside_shell(box_blue, resolution) == False
-assert box_blue.is_inside_shell(box, resolution) == False
+assert box.is_inside_shell(box_blue) == False
+assert box_blue.is_inside_shell(box) == False
 
 model = vm.core.VolumeModel([box_red])
 model.to_step('block.step')